DinhNLHE140992_User_Authorization_Manager

+++ src/java/controller/ManagementPostController.java
@@ -39,24 +39,34 @@ public class ManagementPostController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        int numberPostClick = 1;
-        request.setAttribute("error", null);
-        PostDAO pDAO = new PostDAO();
-        try {
-            int postId = Integer.parseInt(request.getParameter("idDelete"));
-            if (pDAO.removePostByPostId(postId)) {
-                List<PostJoinUser> listPost = new ArrayList<>();
-                listPost = pDAO.getAllPostJoinUser();
-                request.setAttribute("listPost", listPost);
-                request.setAttribute("numberPostClick", numberPostClick);
-                request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+        HttpSession session = request.getSession();
+        User u = (User) session.getAttribute("user");
+        if (u != null) {
+            if (u.getRoleId() == 1 || u.getRoleId() == 4) {
+                int numberPostClick = 1;
+                request.setAttribute("error", null);
+                PostDAO pDAO = new PostDAO();
+                try {
+                    int postId = Integer.parseInt(request.getParameter("idDelete"));
+                    if (pDAO.removePostByPostId(postId)) {
+                        List<PostJoinUser> listPost = new ArrayList<>();
+                        listPost = pDAO.getAllPostJoinUser();
+                        request.setAttribute("listPost", listPost);
+                        request.setAttribute("numberPostClick", numberPostClick);
+                        request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+                    }
+                } catch (Exception e) {
+                    List<PostJoinUser> listPost = new ArrayList<>();
+                    listPost = pDAO.getAllPostJoinUser();
+                    request.setAttribute("listPost", listPost);
+                    request.setAttribute("numberPostClick", numberPostClick);
+                    request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+                }
+            } else {
+                request.getRequestDispatcher("error.jsp").forward(request, response);
             }
-        } catch (Exception e) {
-            List<PostJoinUser> listPost = new ArrayList<>();
-            listPost = pDAO.getAllPostJoinUser();
-            request.setAttribute("listPost", listPost);
-            request.setAttribute("numberPostClick", numberPostClick);
-            request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+        } else {
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
     }
 
@@ -71,45 +81,52 @@ public class ManagementPostController extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String title = ("".equals(request.getParameter("title"))) ? "" : request.getParameter("title");
-        String urlThumbnail = ("".equals(request.getParameter("urlThumbnail"))) ? "" : request.getParameter("urlThumbnail");
-        String urlDetail = ("".equals(request.getParameter("urlDetail"))) ? "" : request.getParameter("urlDetail");
-        String content = ("".equals(request.getParameter("content"))) ? "" : request.getParameter("content");
-        int numberPostClick = 1;
-        PostDAO pDAO = new PostDAO();
-        List<PostJoinUser> listPost = new ArrayList<>();
+        HttpSession session = request.getSession();
+        User u = (User) session.getAttribute("user");
+        if (u != null) {
+            if (u.getRoleId() == 1 || u.getRoleId() == 4) {
+                String title = ("".equals(request.getParameter("title"))) ? "" : request.getParameter("title");
+                String urlThumbnail = ("".equals(request.getParameter("urlThumbnail"))) ? "" : request.getParameter("urlThumbnail");
+                String urlDetail = ("".equals(request.getParameter("urlDetail"))) ? "" : request.getParameter("urlDetail");
+                String content = ("".equals(request.getParameter("content"))) ? "" : request.getParameter("content");
+                int numberPostClick = 1;
+                PostDAO pDAO = new PostDAO();
+                List<PostJoinUser> listPost = new ArrayList<>();
 
-        if (!"".equals(title)) {
-            HttpSession session = request.getSession();
-            User u = (User) session.getAttribute("user");
+                if (!"".equals(title)) {
+                    if (u != null) {
+                        if (pDAO.createPost(u.getUserId(), title, urlThumbnail, urlDetail, content)) {
+                            listPost = pDAO.getAllPostJoinUser();
+                            request.setAttribute("listPost", listPost);
+                            request.setAttribute("error", "1");
+                            request.setAttribute("numberPostClick", numberPostClick);
 
-            if (u != null) {
-                if (pDAO.createPost(u.getUserId(), title, urlThumbnail, urlDetail, content)) {
-                    listPost = pDAO.getAllPostJoinUser();
-                    request.setAttribute("listPost", listPost);
-                    request.setAttribute("error", "1");
-                    request.setAttribute("numberPostClick", numberPostClick);
-                    
-                    try {
-                        SendEmail.sendEmailToSubscriber("[^^] Let check new Post", "New Post: " + title + " [<a href='http://localhost:8080/SWP391.E-BL5_Group2/post'>more...</a>]");
-                    } catch (MessagingException ex) {
-                        Logger.getLogger(ManagementPostController.class.getName()).log(Level.SEVERE, null, ex);
+                            try {
+                                SendEmail.sendEmailToSubscriber("[^^] Let check new Post", "New Post: " + title + " [<a href='http://localhost:8080/SWP391.E-BL5_Group2/post'>more...</a>]");
+                            } catch (MessagingException ex) {
+                                Logger.getLogger(ManagementPostController.class.getName()).log(Level.SEVERE, null, ex);
+                            }
+                            request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+                        } else {
+                            listPost = pDAO.getAllPostJoinUser();
+                            request.setAttribute("listPost", listPost);
+                            request.setAttribute("error", "2");
+                            request.setAttribute("numberPostClick", numberPostClick);
+                            request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+                        }
+                    } else {
+                        listPost = pDAO.getAllPostJoinUser();
+                        request.setAttribute("listPost", listPost);
+                        request.setAttribute("error", "2");
+                        request.setAttribute("numberPostClick", numberPostClick);
+                        request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
                     }
-                    request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
-                } else {
-                    listPost = pDAO.getAllPostJoinUser();
-                    request.setAttribute("listPost", listPost);
-                    request.setAttribute("error", "2");
-                    request.setAttribute("numberPostClick", numberPostClick);
-                    request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
                 }
             } else {
-                listPost = pDAO.getAllPostJoinUser();
-                request.setAttribute("listPost", listPost);
-                request.setAttribute("error", "2");
-                request.setAttribute("numberPostClick", numberPostClick);
-                request.getRequestDispatcher("dashboard/dashboardPost.jsp").forward(request, response);
+                request.getRequestDispatcher("error.jsp").forward(request, response);
             }
+        } else {
+            request.getRequestDispatcher("error.jsp").forward(request, response);
         }
     }
